deploy-requirements.txt
# Deployment Pipeline Dependencies
requests>=2.31.0
google-generativeai>=0.3.0
asyncio-mqtt>=0.13.0
aiohttp>=3.8.0
pydantic>=2.0.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
websockets>=11.0.0
